% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_chart.R
\name{export_chart}
\alias{export_chart}
\title{export_chart}
\usage{
export_chart(
  chart_name,
  file_name,
  include_logo = c(T, F),
  export_width = NULL,
  export_height = NULL,
  export_res = NULL
)
}
\arguments{
\item{chart_name}{Name of ggplot object to export}

\item{file_name}{File path for exporting}

\item{include_logo}{Boolean value - include logo?}

\item{export_width}{Width of exported chart *optional}

\item{export_height}{Height of exported chart *optional}

\item{export_res}{Resolution of exported chart *optional}
}
\description{
This script contains the functions for exporting ggplot charts
with and without the WFA logo.
}
