% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_wfa.R
\name{scale_colour_wfa}
\alias{scale_colour_wfa}
\title{WFA Colour Scale}
\usage{
scale_colour_wfa(palette = "main", discrete = TRUE, reverse = FALSE)
}
\arguments{
\item{palette}{Select palette to use. Defaults to "main".}

\item{discrete}{Defaults to TRUE. Set to FALSE for continuous variable.}

\item{reverse}{Defaults to FALSE. Set to TRUE to reverse the order of the
colour scale.}
}
\description{
This function applies an WFA-themed colour scale to a ggplot object.
}
\examples{
scale_colour_wfa("fuchsia")
scale_colour_wfa(palette = "main", discrete = "TRUE", reverse = "FALSE")
}
\keyword{colour}
\keyword{scale}
