% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{scale_fill_wfa}
\alias{scale_fill_wfa}
\title{scale_fill_wfa}
\usage{
scale_fill_wfa(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of colour palette (e.g. "main", "mixed", "grey")}

\item{discrete}{Boolean value (discrete var = T, continuous var = F)}

\item{reverse}{Boolean value (reverse colour palette = T, don't reverse = F)}

\item{...}{Other optional parameters (see ggplot documentation)}
}
\value{
ggplot fill scale in WFA colours
}
\description{
Generate ggplot fill scale in WFA colours.
}
\examples{
scale_fill_wfa("main", TRUE, FALSE)
scale_fill_wfa("mixed", FALSE, FALSE)
}
