% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wfa.R
\name{theme_wfa}
\alias{theme_wfa}
\title{theme_wfa}
\usage{
theme_wfa()
}
\value{
ggplot theme objects
}
\description{
This script contains the parameters for themeing ggplots in the
WFA style.
}
\examples{
#ggplot2(data = my_data, aes(x = xval, y= yval)) +
#geom_line() +
#theme_wfa()
}
