% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{wfa_cols}
\alias{wfa_cols}
\title{wfa_cols}
\usage{
wfa_cols(...)
}
\arguments{
\item{...}{Name of colour (e.g. "red", "blue")}
}
\value{
Corresponding HEX codes for WFA colours
}
\description{
This function returns the HEX codes for WFA colours.
}
\examples{
wfa_cols("red")
wfa_cols("dark grey")
}
