% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{wfa_pal}
\alias{wfa_pal}
\title{wfa_pal}
\usage{
wfa_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of colour palette (e.g. "main", "mixed", "grey")}

\item{reverse}{Boolean value (reverse colour palette = T, don't reverse = F)}

\item{...}{Other optional parameters (see ggplot documentation)}
}
\value{
WFA colour palette for use with ggplot object
}
\description{
This function returns the WFA colour palettes.
}
\examples{
wfa_pal("main", TRUE)
wfa_pal("mixed", FALSE)
}
