% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_chart.R
\name{export_chart}
\alias{export_chart}
\title{export_chart}
\usage{
export_chart(
  chart_name,
  file_name,
  include_logo = c(T, F),
  export_width = NULL,
  export_height = NULL,
  export_res = NULL
)
}
\arguments{
\item{chart_name}{Name of ggplot object to export}

\item{file_name}{File path for exported chart}

\item{include_logo}{Boolean value (include logo = T, no logo = F)}

\item{export_width}{Width of exported chart (in) *optional}

\item{export_height}{Height of exported chart (in) *optional}

\item{export_res}{Resolution of exported chart (dpi) *optional}
}
\value{
ggplot chart saved as png in root directory
}
\description{
This function exports ggplot charts with and without the WFA
logo.
}
\details{
-----------------------------------------------------------------------------
WFAPLOT - EXPORT CHARTS
export_chart.R
-----------------------------------------------------------------------------
}
\examples{
#export_chart(my_chart, "my_chart.png", TRUE)
#export_chart(my_chart, "my_chart.png", FALSE, 6, 2.5, 300)
}
