% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_wfa.R
\name{scale_fill_wfa}
\alias{scale_fill_wfa}
\title{WFA Fill Scale}
\usage{
scale_fill_wfa(palette = "main", discrete = TRUE, reverse = FALSE)
}
\arguments{
\item{palette}{Select palette to use. Defaults to "main".}

\item{discrete}{Defaults to TRUE. Set to FALSE for continuous variable.}

\item{reverse}{Defaults to FALSE. Set to TRUE to reverse the order of the
fill scale.}
}
\description{
This function applies an WFA-themed fill scale to a ggplot object.
}
\examples{
scale_fill_wfa("fuchsia")
scale_fill_wfa(palette = "main", discrete = "TRUE", reverse = "FALSE")
}
\keyword{fill}
\keyword{scale}
