% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wfa.R
\name{theme_wfa}
\alias{theme_wfa}
\title{WFA Theme}
\usage{
theme_wfa()
}
\description{
This function checks if required fonts are installed and applies the WFA
theme to a ggplot object.
}
\examples{
\donttest{
theme_wfa()
}
}
\keyword{theme}
\keyword{wfa}
