% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfa_cols.R
\name{wfa_cols}
\alias{wfa_cols}
\title{WFA Colours}
\usage{
wfa_cols(cols = "")
}
\arguments{
\item{cols}{Name a colour from the WFA palette to return the HEX code.
Leave argument empty to return all colours.
Choose from red, orange, yellow, lime, green, teal, blue, pink, purple,
indigo, dark grey, and light grey.}
}
\description{
This function returns the HEX codes for WFA colours.
}
\examples{
wfa_cols()
wfa_cols("red")
}
\keyword{color}
\keyword{colour}
