% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfa_pal.R
\name{wfa_pal}
\alias{wfa_pal}
\title{WFA Colour Palettes}
\usage{
wfa_pal(palette = "main", reverse = FALSE)
}
\arguments{
\item{palette}{Select an WFA colour palette. Choose from 'main', 'grey',
'colourful', 'extra_colourful', and 'fuchsia'.}

\item{reverse}{Defaults to FALSE. Set to TRUE to reverse the order of the
colour palette}
}
\description{
This function returns an WFA colour palette.
}
\examples{
wfa_pal()
wfa_pal("extra_colourful", reverse = TRUE)
}
\keyword{palette}
